pyproject.toml
[tool.poetry]
name = "tgfp-web"
version = "2.4.3"
description = "The Great Football Pool flask web site"
authors = ["John Sturgeon <john.sturgeon@me.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
gunicorn = "^21.2.0"
Flask-Discord = "^0.1.69"
blinker = "^1.6.2"
tgfp-lib = "^2.1.1"
Flask-DebugToolbar = "^0.13.1"
setuptools = "^68.2.2"
pylint = "^2.17.5"
fastapi = "^0.103.1"
prefect = "^2.13.1"
greenlet = "^2.0.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
